requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-Login>=0.6
openpyxl>=3.0
matplotlib>=3.5
numpy>=1.20
streamlit>=1.28
pandas>=1.5
